+++ scripts/build_ext_all.sh
@@ -13,11 +13,11 @@ declare -A plats
 
 plats["linux/x86_64"]="native"
 #plats["openwrt/arm_cortex-a9_neon"]="imx6_exports.sh"
-plats["openwrt/arm_cortex-a9_neon"]="tg3030_exports.sh"
-plats["openwrt/mips_24kc"]="mips_24kc.sh"
-plats["openwrt/x86_64"]="x86_64_glibc.sh"
+plats["openwrt/17.01/arm_cortex-a9_neon"]="tg3030_exports.sh"
+plats["openwrt/18.06/mips_24kc"]="mips_24kc.sh"
+plats["openwrt/18.06/x86_64"]="x86_64_glibc.sh"
 #plats["openwrt/aarch64_cortex-a53"]="bp3plus_exports.sh"
-plats["openwrt/arm_cortex-a7_neon-vfpv4"]="sunxi_a7.sh"
+plats["openwrt/19.07/arm_cortex-a7_neon-vfpv4"]="sunxi_a7.sh"
 # plats["android/arm"]="android_arm.sh"
 
 for item in "${!plats[@]}"; 
@@ -25,5 +25,6 @@ do
 	if [ "$BUILD_LIB" == "opcua" ]; then
 		bash $SCRIPTPATH/build_ext_opcua.sh $item ${plats[$item]} $CUR_DIR/../..
 	fi
+	mkdir -p $SCRIPTPATH/../feeds/prebuild_exts/$item/
 	bash $SCRIPTPATH/build_ext.sh $item ${plats[$item]} $SCRIPTPATH/../feeds/prebuild_exts/$item/
 done