+++ scripts/make_release.sh
@@ -53,7 +53,7 @@ cp -rL lualib/resty __install/lualib/resty
 ./feeds/example_apps/release.sh ./scripts/release_app.sh
 
 # Validate platform name
-PLAT_NAMES="linux/x86_64 openwrt/arm_cortex-a9_neon openwrt/mips_24kc openwrt/x86_64 openwrt/aarch64_cortex-a53 openwrt/arm_cortex-a7_neon-vfpv4"
+PLAT_NAMES="linux/x86_64 openwrt/17.01/arm_cortex-a9_neon openwrt/18.06/x86_64 openwrt/19.07/arm_cortex-a7_neon-vfpv4"
 
 # Release Extensions
 for plat in $PLAT_NAMES; do